refactor(config-rewrite): remove @Type as it's replaced by @OptionType.TypeTarget, move common classes to api root
